requirements.txt
Flask==2.0.3
Pandas
numpy
Pillow
gunicorn == 20.0.4
tensorflow-cpu
keras
h5py
efficientnet
